Add PR and Rf pages
