+++ src/components/RefNameReports.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import Reports from "./Reports";
+
+export default class RefNameReports extends React.Component {
+	render() {
+		return (
+				<Reports event='push' />
+		);
+	}
+}
